requirements.txt
pydantic
requests
ttkthemes
tkintermapviews